Remove extra classes from RegistrationViewModel
